turtle mode doesn't work for online user, obviously
